mild refactoring
